add init suadmin and login

+++ app/users/users.js
@@ -2,19 +2,31 @@
 
 const crypto = require('crypto');
 
-const createNamespace = require('continuation-local-storage').createNamespace;
-const getNamespace = require('continuation-local-storage').getNamespace;
-const uuid = require('uuidv4');
+const getNamespace = require('cls-hooked').getNamespace;
+const uuidV4 = require('uuidv4');
 
 const dbConnector = require('../db/db-connector');
 
 const USERS_COLL_NAME = 'users';
 const SESSION_EXPIRATION_DELAY_HOURS = 2;
 
-const sessions = new Map();
-const setSessionCls = createNamespace('sessions');
-const getSessionCls = getNamespace('sessions');
+const SUPER_ADMIN_LOGIN = 'admin';
 
+const sessions = new Map();
+
+const secondsInMinute= 60;
+const milliseconds = 1000;
+setInterval(() => {
+	const deprecatedSessionIds = [];
+	sessions.forEach((session, sessionId) => {
+		if (!session.isValid) {
+			deprecatedSessionIds.push(sessionId);
+		}
+	});
+	deprecatedSessionIds.forEach(deprecatedSessionId => {
+		sessions.delete(deprecatedSessionId);
+	});
+}, secondsInMinute * milliseconds);
 
 class Session {
 	constructor(user) {
@@ -37,11 +49,11 @@ class Session {
 }
 
 class User {
-	constructor({uuid = uuid(), login = '', pass = '', email = '', firstName = '', lastName = '', readOnly = true}) {
+	constructor({uuid = uuidV4(), login = '', pass = '', email = '', firstName = '', lastName = '', readOnly = true, hashPass = false}) {
 		// TODO some params validations
 		this._id = uuid;
 		this.login = login;
-		this.pass = this.hashPass(pass);
+		this.pass = hashPass ? User.hashPass(pass) : pass;
 		this.email = email;
 		this.firstName = firstName;
 		this.lastName = lastName;
@@ -54,6 +66,13 @@ class User {
 		return hash.digest('hex');
 	}
 
+	static get superAdminLogin() {
+		return SUPER_ADMIN_LOGIN;
+	}
+
+	get isSuperAdmin() {
+		return this.login === User.superAdminLogin;
+	}
 }
 
 async function getUsers() {
@@ -61,11 +80,21 @@ async function getUsers() {
 	const cursor = await coll.find();
 	const itemsCount = await cursor.count();
 	if (itemsCount > 0){
-		return cursor.toArray().map(rawUser => new User(rawUser));
+		return (await cursor.toArray()).map(rawUser => new User(rawUser));
 	}
 	const err = new Error('No user found in database');
 	err.code = 'ENOUSERFOUND';
 	throw err;
+}
+
+async function getSuperAdmin() {
+	const coll = await dbConnector.getCollection(USERS_COLL_NAME);
+	const cursor = await coll.find({login: User.superAdminLogin});
+	const itemsCount = await cursor.count();
+	if (itemsCount > 0){
+		return new User(await cursor.next());
+	}
+	return null;
 }
 
 async function addUser(userProps) {
@@ -88,20 +117,19 @@ async function addUser(userProps) {
 }
 
 async function authenticate(login, pass, sessId) {
-	const users = await getUsers();
-
-	if (!users.find(userItem => userItem.login === login)) {
-		throw new Error('User not found');
-	}
-
-	if (sessions.has(sessId)) {
+	const sessionCls = getNamespace('sessions');
+	if (sessId && sessions.has(sessId)) {
 		const session = sessions.get(sessId);
 		if (session.isValid) {
-			setSessionCls.set('user', session.user);
-			setSessionCls.set('sessId', sessId);
+			sessionCls.set('user', session.user);
+			sessionCls.set('sessId', sessId);
 			return true;
 		}
 		session.delete(sessId);
+	}
+
+	const users = await getUsers();
+	if (!users.find(userItem => userItem.login === login)) {
 		return false;
 	}
 
@@ -111,23 +139,29 @@ async function authenticate(login, pass, sessId) {
 	});
 
 	if (foundUser) {
-		const newSessId = uuid();
+		const newSessId = uuidV4();
 		sessions.set(newSessId, new Session(foundUser));
-		setSessionCls.set('sessId', newSessId);
-		return true;
+		sessionCls.set('user', foundUser);
+		sessionCls.set('sessId', newSessId);
+		return true
 	}
 
-	return false;
+	return false
 }
+
 
 module.exports = {
 	authenticate,
-	getSessId() {
-		return getSessionCls.get('sessId');
-	},
-	getCurrentUser() {
-		return getSessionCls.get('user');
-	},
 	getUsers,
-	addUser
+	getSuperAdmin,
+	addUser,
+	User,
+	getCurrentUser() {
+		const sessionCls = getNamespace('sessions');
+		return sessionCls.get('user');
+	},
+	getSessId() {
+		const sessionCls = getNamespace('sessions');
+		return sessionCls.get('sessId');
+	}
 };
